+++ crossbeam-channel/src/lib.rs
@@ -348,7 +348,7 @@
 #![warn(missing_debug_implementations)]
 
 extern crate crossbeam_utils;
-extern crate smallvec;
+extern crate maybe_uninit;
 
 mod channel;
 mod context;
@@ -359,6 +359,13 @@ mod select;
 mod select_macro;
 mod utils;
 mod waker;
+
+/// Crate internals used by the `select!` macro.
+#[doc(hidden)]
+pub mod internal {
+    pub use select::SelectHandle;
+    pub use select::{select, select_timeout, try_select};
+}
 
 pub use channel::{after, never, tick};
 pub use channel::{bounded, unbounded};