+++ crossbeam-utils/src/atomic/atomic_cell.rs
@@ -2,10 +2,12 @@ use core::cell::UnsafeCell;
 use core::fmt;
 use core::mem;
 use core::ptr;
-use core::slice;
-use core::sync::atomic::{self, AtomicBool, AtomicUsize, Ordering};
+use core::sync::atomic::{self, AtomicBool, Ordering};
 
-use Backoff;
+#[cfg(feature = "std")]
+use std::panic::{RefUnwindSafe, UnwindSafe};
+
+use super::seq_lock::SeqLock;
 
 /// A thread-safe mutable memory location.
 ///
@@ -21,7 +23,8 @@ use Backoff;
 /// [`AtomicCell::<T>::is_lock_free()`]: struct.AtomicCell.html#method.is_lock_free
 /// [`Acquire`]: https://doc.rust-lang.org/std/sync/atomic/enum.Ordering.html#variant.Acquire
 /// [`Release`]: https://doc.rust-lang.org/std/sync/atomic/enum.Ordering.html#variant.Release
-pub struct AtomicCell<T> {
+#[repr(transparent)]
+pub struct AtomicCell<T: ?Sized> {
     /// The inner value.
     ///
     /// If this value can be transmuted into a primitive atomic type, it will be treated as such.
@@ -33,6 +36,11 @@ pub struct AtomicCell<T> {
 unsafe impl<T: Send> Send for AtomicCell<T> {}
 unsafe impl<T: Send> Sync for AtomicCell<T> {}
 
+#[cfg(feature = "std")]
+impl<T> UnwindSafe for AtomicCell<T> {}
+#[cfg(feature = "std")]
+impl<T> RefUnwindSafe for AtomicCell<T> {}
+
 impl<T> AtomicCell<T> {
     /// Creates a new atomic cell initialized with `val`.
     ///
@@ -43,28 +51,27 @@ impl<T> AtomicCell<T> {
     ///
     /// let a = AtomicCell::new(7);
     /// ```
+    #[cfg(not(has_min_const_fn))]
     pub fn new(val: T) -> AtomicCell<T> {
         AtomicCell {
             value: UnsafeCell::new(val),
         }
     }
 
-    /// Returns a mutable reference to the inner value.
+    /// Creates a new atomic cell initialized with `val`.
     ///
     /// # Examples
     ///
     /// ```
     /// use crossbeam_utils::atomic::AtomicCell;
     ///
-    /// let mut a = AtomicCell::new(7);
-    /// *a.get_mut() += 1;
-    ///
-    /// assert_eq!(a.load(), 8);
+    /// let a = AtomicCell::new(7);
     /// ```
-    #[doc(hidden)]
-    #[deprecated(note = "this method is unsound and will be removed in the next release")]
-    pub fn get_mut(&mut self) -> &mut T {
-        unsafe { &mut *self.value.get() }
+    #[cfg(has_min_const_fn)]
+    pub const fn new(val: T) -> AtomicCell<T> {
+        AtomicCell {
+            value: UnsafeCell::new(val),
+        }
     }
 
     /// Unwraps the atomic cell and returns its inner value.
@@ -156,6 +163,61 @@ impl<T> AtomicCell<T> {
     }
 }
 
+impl<T: ?Sized> AtomicCell<T> {
+    /// Returns a raw pointer to the underlying data in this atomic cell.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use crossbeam_utils::atomic::AtomicCell;
+    ///
+    /// let mut a = AtomicCell::new(5);
+    ///
+    /// let ptr = a.as_ptr();
+    /// ```
+    #[inline]
+    pub fn as_ptr(&self) -> *mut T {
+        self.value.get()
+    }
+
+    /// Returns a mutable reference to the inner value.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use crossbeam_utils::atomic::AtomicCell;
+    ///
+    /// let mut a = AtomicCell::new(7);
+    /// *a.get_mut() += 1;
+    ///
+    /// assert_eq!(a.load(), 8);
+    /// ```
+    #[doc(hidden)]
+    #[deprecated(note = "this method is unsound and will be removed in the next release")]
+    pub fn get_mut(&mut self) -> &mut T {
+        unsafe { &mut *self.value.get() }
+    }
+}
+
+impl<T: Default> AtomicCell<T> {
+    /// Takes the value of the atomic cell, leaving `Default::default()` in its place.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use crossbeam_utils::atomic::AtomicCell;
+    ///
+    /// let a = AtomicCell::new(5);
+    /// let five = a.take();
+    ///
+    /// assert_eq!(five, 5);
+    /// assert_eq!(a.into_inner(), 0);
+    /// ```
+    pub fn take(&self) -> T {
+        self.swap(Default::default())
+    }
+}
+
 impl<T: Copy> AtomicCell<T> {
     /// Loads a value.
     ///
@@ -217,23 +279,8 @@ impl<T: Copy + Eq> AtomicCell<T> {
     /// assert_eq!(a.compare_exchange(1, 2), Ok(1));
     /// assert_eq!(a.load(), 2);
     /// ```
-    pub fn compare_exchange(&self, mut current: T, new: T) -> Result<T, T> {
-        loop {
-            match unsafe { atomic_compare_exchange_weak(self.value.get(), current, new) } {
-                Ok(_) => return Ok(current),
-                Err(previous) => {
-                    if previous != current {
-                        return Err(previous);
-                    }
-
-                    // The compare-exchange operation has failed and didn't store `new`. The
-                    // failure is either spurious, or `previous` was semantically equal to
-                    // `current` but not byte-equal. Let's retry with `previous` as the new
-                    // `current`.
-                    current = previous;
-                }
-            }
-        }
+    pub fn compare_exchange(&self, current: T, new: T) -> Result<T, T> {
+        unsafe { atomic_compare_exchange_weak(self.value.get(), current, new) }
     }
 }
 
@@ -472,37 +519,28 @@ macro_rules! impl_arithmetic {
             }
         }
     };
-    ($t:ty, $size:tt, $atomic:ty, $example:tt) => {
-        #[cfg(target_has_atomic = $size)]
-        impl_arithmetic!($t, $atomic, $example);
-    };
 }
 
-cfg_if! {
-    if #[cfg(feature = "nightly")] {
-        impl_arithmetic!(u8, "8", atomic::AtomicU8, "let a = AtomicCell::new(7u8);");
-        impl_arithmetic!(i8, "8", atomic::AtomicI8, "let a = AtomicCell::new(7i8);");
-        impl_arithmetic!(u16, "16", atomic::AtomicU16, "let a = AtomicCell::new(7u16);");
-        impl_arithmetic!(i16, "16", atomic::AtomicI16, "let a = AtomicCell::new(7i16);");
-        impl_arithmetic!(u32, "32", atomic::AtomicU32, "let a = AtomicCell::new(7u32);");
-        impl_arithmetic!(i32, "32", atomic::AtomicI32, "let a = AtomicCell::new(7i32);");
-        impl_arithmetic!(u64, "64", atomic::AtomicU64, "let a = AtomicCell::new(7u64);");
-        impl_arithmetic!(i64, "64", atomic::AtomicI64, "let a = AtomicCell::new(7i64);");
-        impl_arithmetic!(u128, "let a = AtomicCell::new(7u128);");
-        impl_arithmetic!(i128, "let a = AtomicCell::new(7i128);");
-    } else {
-        impl_arithmetic!(u8, "let a = AtomicCell::new(7u8);");
-        impl_arithmetic!(i8, "let a = AtomicCell::new(7i8);");
-        impl_arithmetic!(u16, "let a = AtomicCell::new(7u16);");
-        impl_arithmetic!(i16, "let a = AtomicCell::new(7i16);");
-        impl_arithmetic!(u32, "let a = AtomicCell::new(7u32);");
-        impl_arithmetic!(i32, "let a = AtomicCell::new(7i32);");
-        impl_arithmetic!(u64, "let a = AtomicCell::new(7u64);");
-        impl_arithmetic!(i64, "let a = AtomicCell::new(7i64);");
-        impl_arithmetic!(u128, "let a = AtomicCell::new(7u128);");
-        impl_arithmetic!(i128, "let a = AtomicCell::new(7i128);");
-    }
-}
+#[cfg(has_atomic_u8)]
+impl_arithmetic!(u8, atomic::AtomicU8, "let a = AtomicCell::new(7u8);");
+#[cfg(has_atomic_u8)]
+impl_arithmetic!(i8, atomic::AtomicI8, "let a = AtomicCell::new(7i8);");
+#[cfg(has_atomic_u16)]
+impl_arithmetic!(u16, atomic::AtomicU16, "let a = AtomicCell::new(7u16);");
+#[cfg(has_atomic_u16)]
+impl_arithmetic!(i16, atomic::AtomicI16, "let a = AtomicCell::new(7i16);");
+#[cfg(has_atomic_u32)]
+impl_arithmetic!(u32, atomic::AtomicU32, "let a = AtomicCell::new(7u32);");
+#[cfg(has_atomic_u32)]
+impl_arithmetic!(i32, atomic::AtomicI32, "let a = AtomicCell::new(7i32);");
+#[cfg(has_atomic_u64)]
+impl_arithmetic!(u64, atomic::AtomicU64, "let a = AtomicCell::new(7u64);");
+#[cfg(has_atomic_u64)]
+impl_arithmetic!(i64, atomic::AtomicI64, "let a = AtomicCell::new(7i64);");
+#[cfg(has_atomic_u128)]
+impl_arithmetic!(u128, atomic::AtomicU128, "let a = AtomicCell::new(7u128);");
+#[cfg(has_atomic_u128)]
+impl_arithmetic!(i128, atomic::AtomicI128, "let  a = AtomicCell::new(7i128);");
 
 impl_arithmetic!(
     usize,
@@ -594,100 +632,10 @@ impl<T: Copy + fmt::Debug> fmt::Debug for AtomicCell<T> {
     }
 }
 
-/// Returns `true` if the two values are equal byte-for-byte.
-fn byte_eq<T>(a: &T, b: &T) -> bool {
-    unsafe {
-        let a = slice::from_raw_parts(a as *const _ as *const u8, mem::size_of::<T>());
-        let b = slice::from_raw_parts(b as *const _ as *const u8, mem::size_of::<T>());
-        a == b
-    }
-}
-
 /// Returns `true` if values of type `A` can be transmuted into values of type `B`.
 fn can_transmute<A, B>() -> bool {
     // Sizes must be equal, but alignment of `A` must be greater or equal than that of `B`.
     mem::size_of::<A>() == mem::size_of::<B>() && mem::align_of::<A>() >= mem::align_of::<B>()
-}
-
-/// A simple stamped lock.
-struct Lock {
-    /// The current state of the lock.
-    ///
-    /// All bits except the least significant one hold the current stamp. When locked, the state
-    /// equals 1 and doesn't contain a valid stamp.
-    state: AtomicUsize,
-}
-
-impl Lock {
-    /// If not locked, returns the current stamp.
-    ///
-    /// This method should be called before optimistic reads.
-    #[inline]
-    fn optimistic_read(&self) -> Option<usize> {
-        let state = self.state.load(Ordering::Acquire);
-        if state == 1 {
-            None
-        } else {
-            Some(state)
-        }
-    }
-
-    /// Returns `true` if the current stamp is equal to `stamp`.
-    ///
-    /// This method should be called after optimistic reads to check whether they are valid. The
-    /// argument `stamp` should correspond to the one returned by method `optimistic_read`.
-    #[inline]
-    fn validate_read(&self, stamp: usize) -> bool {
-        atomic::fence(Ordering::Acquire);
-        self.state.load(Ordering::Relaxed) == stamp
-    }
-
-    /// Grabs the lock for writing.
-    #[inline]
-    fn write(&'static self) -> WriteGuard {
-        let backoff = Backoff::new();
-        loop {
-            let previous = self.state.swap(1, Ordering::Acquire);
-
-            if previous != 1 {
-                atomic::fence(Ordering::Release);
-
-                return WriteGuard {
-                    lock: self,
-                    state: previous,
-                };
-            }
-
-            backoff.snooze();
-        }
-    }
-}
-
-/// A RAII guard that releases the lock and increments the stamp when dropped.
-struct WriteGuard {
-    /// The parent lock.
-    lock: &'static Lock,
-
-    /// The stamp before locking.
-    state: usize,
-}
-
-impl WriteGuard {
-    /// Releases the lock without incrementing the stamp.
-    #[inline]
-    fn abort(self) {
-        self.lock.state.store(self.state, Ordering::Release);
-    }
-}
-
-impl Drop for WriteGuard {
-    #[inline]
-    fn drop(&mut self) {
-        // Release the lock and increment the stamp.
-        self.lock
-            .state
-            .store(self.state.wrapping_add(2), Ordering::Release);
-    }
 }
 
 /// Returns a reference to the global lock associated with the `AtomicCell` at address `addr`.
@@ -700,7 +648,7 @@ impl Drop for WriteGuard {
 /// scalability.
 #[inline]
 #[must_use]
-fn lock(addr: usize) -> &'static Lock {
+fn lock(addr: usize) -> &'static SeqLock {
     // The number of locks is a prime number because we want to make sure `addr % LEN` gets
     // dispersed across all locks.
     //
@@ -725,10 +673,9 @@ fn lock(addr: usize) -> &'static Lock {
     // In order to protect from such cases, we simply choose a large prime number for `LEN`.
     const LEN: usize = 97;
 
-    const L: Lock = Lock {
-        state: AtomicUsize::new(0),
-    };
-    static LOCKS: [Lock; LEN] = [
+    const L: SeqLock = SeqLock::INIT;
+
+    static LOCKS: [SeqLock; LEN] = [
         L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L,
         L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L,
         L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L, L,
@@ -785,17 +732,14 @@ macro_rules! atomic {
             atomic!(@check, $t, AtomicUnit, $a, $atomic_op);
             atomic!(@check, $t, atomic::AtomicUsize, $a, $atomic_op);
 
-            #[cfg(feature = "nightly")]
-            {
-                #[cfg(target_has_atomic = "8")]
-                atomic!(@check, $t, atomic::AtomicU8, $a, $atomic_op);
-                #[cfg(target_has_atomic = "16")]
-                atomic!(@check, $t, atomic::AtomicU16, $a, $atomic_op);
-                #[cfg(target_has_atomic = "32")]
-                atomic!(@check, $t, atomic::AtomicU32, $a, $atomic_op);
-                #[cfg(target_has_atomic = "64")]
-                atomic!(@check, $t, atomic::AtomicU64, $a, $atomic_op);
-            }
+            #[cfg(has_atomic_u8)]
+            atomic!(@check, $t, atomic::AtomicU8, $a, $atomic_op);
+            #[cfg(has_atomic_u16)]
+            atomic!(@check, $t, atomic::AtomicU16, $a, $atomic_op);
+            #[cfg(has_atomic_u32)]
+            atomic!(@check, $t, atomic::AtomicU32, $a, $atomic_op);
+            #[cfg(has_atomic_u64)]
+            atomic!(@check, $t, atomic::AtomicU64, $a, $atomic_op);
 
             break $fallback_op;
         }
@@ -857,13 +801,12 @@ unsafe fn atomic_store<T>(dst: *mut T, val: T) {
         T, a,
         {
             a = &*(dst as *const _ as *const _);
-            let res = a.store(mem::transmute_copy(&val), Ordering::Release);
+            a.store(mem::transmute_copy(&val), Ordering::Release);
             mem::forget(val);
-            res
         },
         {
             let _guard = lock(dst as usize).write();
-            ptr::write(dst, val)
+            ptr::write(dst, val);
         }
     }
 }
@@ -895,29 +838,46 @@ unsafe fn atomic_swap<T>(dst: *mut T, val: T) -> T {
 ///
 /// This operation uses the `AcqRel` ordering. If possible, an atomic instructions is used, and a
 /// global lock otherwise.
-unsafe fn atomic_compare_exchange_weak<T>(dst: *mut T, current: T, new: T) -> Result<T, T>
+unsafe fn atomic_compare_exchange_weak<T>(dst: *mut T, mut current: T, new: T) -> Result<T, T>
 where
-    T: Copy,
+    T: Copy + Eq,
 {
     atomic! {
         T, a,
         {
             a = &*(dst as *const _ as *const _);
-            let res = a.compare_exchange_weak(
-                mem::transmute_copy(&current),
-                mem::transmute_copy(&new),
-                Ordering::AcqRel,
-                Ordering::Acquire,
-            );
-            match res {
-                Ok(v) => Ok(mem::transmute_copy(&v)),
-                Err(v) => Err(mem::transmute_copy(&v)),
+            let mut current_raw = mem::transmute_copy(&current);
+            let new_raw = mem::transmute_copy(&new);
+
+            loop {
+                match a.compare_exchange_weak(
+                    current_raw,
+                    new_raw,
+                    Ordering::AcqRel,
+                    Ordering::Acquire,
+                ) {
+                    Ok(_) => break Ok(current),
+                    Err(previous_raw) => {
+                        let previous = mem::transmute_copy(&previous_raw);
+
+                        if !T::eq(&previous, &current) {
+                            break Err(previous);
+                        }
+
+                        // The compare-exchange operation has failed and didn't store `new`. The
+                        // failure is either spurious, or `previous` was semantically equal to
+                        // `current` but not byte-equal. Let's retry with `previous` as the new
+                        // `current`.
+                        current = previous;
+                        current_raw = previous_raw;
+                    }
+                }
             }
         },
         {
             let guard = lock(dst as usize).write();
 
-            if byte_eq(&*dst, &current) {
+            if T::eq(&*dst, &current) {
                 Ok(ptr::replace(dst, new))
             } else {
                 let val = ptr::read(dst);